feat(database): removed database connection from the library itself but instead we let people connect to their favorite connection
