+++ package.json
@@ -21,12 +21,10 @@
     "xmigrate": "./dist/main.js"
   },
   "dependencies": {
-    "@rxdi/core": "^0.7.37",
+    "@rxdi/core": "^0.7.178",
     "chalk": "2.4.2",
     "esbuild": "^0.20.2",
-    "esm": "3.2.25",
-    "mongodb": "3.3.3",
-    "mongoose": "5.7.6"
+    "esm": "3.2.25"
   },
   "devDependencies": {
     "@babel/core": "^7.10.2",
@@ -48,6 +46,10 @@
     "prettier": "^2.0.4",
     "ts-jest": "^24.0.2",
     "typescript": "^3.5.3"
+  },
+  "peerDependencies": {
+    "mongodb": "*",
+    "mongoose": "*"
   },
   "jest": {
     "testEnvironment": "node",
