+++ src/services/index.ts
@@ -1,4 +1,3 @@
-export * from './database/database.service';
 export * from './generic-runner/generic-runner.service';
 export * from './migration/migration.service';
 export * from './migrations-resolver/migrations-resolver.service';
